+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASED
 
+* switch build tool to rollup
+
 ## v0.10.2
 
 * fix form validation error for detached elements